Added Edit Workouts
I had to make a few changes to other files to be able to permanently delete workouts.

In the workoutData.ts I only did one change:

I had to make it so saveWorkouts() accepted a parameter so other files like timeScreen.tsx to save workout updates

In the timeScreen.tsx I just added something to make it so it would save updated data.

In the index.tsx for the entire project

I added a few things like a const to render workouts live

I also had to use useFocusEffect in the loadWorkouts so that the homescreen always shows workouts that exist

I also changed Object.values(workoutDetails) to
data={Object.entries(workouts)} because the old one would not update after I made changes.

Also there are a bunch of stuff that was added as the normal way to run expo was not working so I used tunnel. I think the reason this was happening is because of my WiFi.